Backend/requirements.txt
# requirements.txt
# """
# Validated for Python 3.11.9 and Django 4.2 LTS

# Policy:
#  - Keep Django at 4.2 LTS (fully supports Python 3.11)
#  - Keep DRF at 3.14.0 to avoid drf-yasg compatibility issues (drf-yasg often lags behind DRF releases)
#  - Upgrade / adjust packages where Python 3.11 wheels are confirmed stable
#  - Unpin the performance / native build sensitive vision stack (numpy, dlib, face-recognition) so pip can pick
# 	 appropriate wheels for the local platform & Python 3.11. If you need deterministic builds later,
# 	 capture the resolved versions with:  pip freeze > requirements.lock
#  - If further issues arise with dlib / face-recognition on Windows, install Visual C++ Build Tools or
# 	 prefer pre-built wheels from reputable sources.
# """

# --- Core Django & API ---
Django==4.2.16
djangorestframework==3.14.0          # pinned for drf-yasg compatibility
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-jsonapi==6.0.0    # matches DRF 3.14 series

# --- Authentication & Security ---
djangorestframework-simplejwt==5.3.1
cryptography==43.0.1
django-ratelimit==4.1.0
decorator==5.1.1

# --- Face Recognition & Image Processing ---
# IMPORTANT:
#  * face-recognition 1.3.0 currently requires dlib < 20.0.0 (>=19.7)
#  * Latest dlib (20.x) causes dependency conflict; pin to 19.24.2
#  * numpy 1.26.4 is last 1.x line with very wide 3rd‑party wheel support and works with dlib 19.24.2
#  * If you later upgrade to numpy>=2, re-test dlib & face-recognition (may fail)
#  * If building dlib from source on Windows fails, install Build Tools & CMake or use a prebuilt wheel
numpy==1.26.4
dlib==19.24.2
face-recognition==1.3.0
opencv-python==4.9.0.80               # Confirmed cp311 wheel
Pillow==10.3.0                        # cp311 wheel
django-imagekit==5.0.0
pilkit==3.0
base64io==1.0.3

# --- Database ---
# NOTE: psycopg2-binary is for ease of installation.
# For production, plan migration to the 'psycopg' (v3) package.
psycopg2-binary==2.9.9

# --- Asynchronous Tasks ---
celery==5.3.6                         # 5.4+ optional; keep 5.3.x stability
redis==5.0.1
django-redis==5.4.0

# --- API Documentation (Swagger/OpenAPI) ---
# NOTE: coreapi, coreschema, and itypes have been removed as they are
# no longer required by modern versions of drf-yasg.
drf-yasg==1.21.7                     # works with DRF 3.14
uritemplate==4.1.1
ruamel.yaml==0.18.6
inflection==0.5.1

# --- Environment & File Handling ---
django-environ==0.11.2
python-magic-bin==0.4.14
filetype==1.2.0

# --- Utilities ---
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2024.1
arrow==1.3.0

# --- Email ---
# NOTE: Previous pin 0.3.4 was invalid (latest available: 0.3.3). Pin to 0.3.3 for reproducibility.
django-email-verification==0.3.3
html2text==2020.1.16

# --- Production Server ---
gunicorn==21.2.0
whitenoise==6.6.0

# --- Development & Debugging Tools ---
django-debug-toolbar==4.4.2
django-extensions==3.2.3
ipython==8.26.0
django-silk==5.1.0
colorlog==6.8.2
sentry-sdk==1.45.0

# --- Testing ---
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
